test: split GitHub workflow tooling checks
